Add the QMX iOS app: live MM/ML menu access, profiles, simulator
ios/pocket-cat-qmx mirrors the FT-891 app for the QRP Labs QMX:
- Operate: frequency, the QMX's four MD modes + Q1 sideband, RIT,
  split, keyer speed, PTT, and meters (S-meter dB, measured power,
  SWR, AGC)
- Menu: the radio's ENTIRE configuration tree discovered live over the
  QMX Menu Manager (MM/ML) — typed editors per item kind, grid pages
  with per-band columns, and written explanations (QMXMenuNotes)
- Profiles: capture every menu leaf + operating state to .qmxjson in
  iCloud Drive; diff and apply with read-back verification
- QMXSimTransport: full simulator incl. an MM/ML tree, so the app and
  its 25 tests run with no hardware

CATBridgeKit corrections against the real QMX CAT manual (1_02_006):
PC is a GET-only power meter in tenths of a watt (power becomes Double
watts everywhere), SM reads dB, MD accepts only 3/6/7/9. The qmx-cat.md
reference and both radio simulators now match the manual, and the
FT-891 app adapts to the Double power type.

+++ Sources/CATBridgeCore/CAT/CATTypes.swift
@@ -126,8 +126,9 @@ public enum CATValue: Sendable, Equatable {
     case ptt(Bool)
     case sMeter(Int)
     case info(RigInfo)
-    /// RF output power, watts.
-    case power(Int)
+    /// RF output power, watts. Double because the QMX reports measured
+    /// output in tenths of a watt (`PC45;` = 4.5 W — cat_1_02_006).
+    case power(Double)
     case setting(RigSetting, Int)
     /// A menu item (Yaesu `EX`): number as sent, value as the raw digit
     /// string the radio returned.
